% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dists.R
\name{beta_mu}
\alias{beta_mu}
\alias{beta_sd}
\title{Beta Distribution Parameter - mu & sd}
\usage{
beta_mu(alpha, beta)

beta_sd(alpha, beta)
}
\arguments{
\item{alpha}{alpha parameter of a beta distribution}

\item{beta}{beta parameter of a beta distribution}
}
\description{
Calculate mean from alpha and beta parameters of a
beta distribution

Calculate standard deviation from alpha and beta parameters of a
beta distribution
}
