% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dists.R
\name{get_beta_parameters}
\alias{get_beta_parameters}
\title{Get beta distribution parameters}
\usage{
get_beta_parameters(population_avg, population_var)
}
\arguments{
\item{population_avg}{the population average, a length 1 numeric value
describing the average (mean) value of the population}

\item{population_var}{the population variance, a length 1 numeric value
describing the variance of the population}
}
\description{
Based on the population average and variance, calculate the alpha and beta
parameters that describe the beta distribution.
}
\details{
The beta distribution should be used when the distribution being predicting
is a value between 0 and 1
}
\examples{

## Set up the data

pop_avg <- lahman_batting_2010 \%>\%
  filter(AB >= quantile(AB, probs = 0.75)) \%>\%
  pull(batting_avg) \%>\%
  mean()
pop_var <- lahman_batting_2010 \%>\%
  filter(AB >= quantile(AB, probs = 0.75)) \%>\%
  pull(batting_avg) \%>\%
  var()

pop_avg
pop_var
sqrt(pop_var)

## Using the function

batting_avg_params <- get_beta_parameters(
  population_avg = pop_avg,
  population_var = pop_var
)

batting_avg_params

}
