% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lahman_batting_2010}
\alias{lahman_batting_2010}
\title{Player Batting}
\format{
\describe{
\item{playerID}{Player Number}
\item{yearID}{Year}
\item{AB}{At Bats}
\item{H}{Hits}
\item{batting_avg}{batting average}
}
}
\usage{
lahman_batting_2010
}
\description{
Subset of batting data from the {lahman} package to include only
batters post 2010 and derive batting average
}
\keyword{datasets}
