% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lahman_team_runs_2010}
\alias{lahman_team_runs_2010}
\title{Team Runs}
\format{
\describe{
\item{yearID}{Year}
\item{teamID}{Team}
\item{G}{Games}
\item{R}{Runs}
\item{runs_per_game}{Runs per game}
}
}
\usage{
lahman_team_runs_2010
}
\description{
Subset of team level runs data from the {lahman} package to
include only post 2010 and derive runs per game
}
\keyword{datasets}
