% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dists.R
\name{posterior_beta_mu}
\alias{posterior_beta_mu}
\alias{posterior_beta_sd}
\title{Calculate Posterior mu based on Beta distribution}
\usage{
posterior_beta_mu(prior_alpha, prior_beta, n, success)

posterior_beta_sd(prior_alpha, prior_beta, n, success)
}
\arguments{
\item{prior_alpha}{beta distribution alpha parameter from population}

\item{prior_beta}{beta distribution beta parameter from population}

\item{n}{numeric value representing number of observations}

\item{success}{numeric value representing the number of success (positive) observations}
}
\description{
Calculate posterior mu and sd given priors, observations, and n success to
calculate
}
