% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dists.R
\name{get_gamma_parameters}
\alias{get_gamma_parameters}
\alias{gamma_mu}
\alias{gamma_sd}
\alias{posterior_gamma_mu}
\alias{posterior_gamma_sd}
\title{Gamma Distribution}
\usage{
get_gamma_parameters(population_avg, population_var)

gamma_mu(alpha, beta)

gamma_sd(alpha, beta)

posterior_gamma_mu(alpha, beta, n, successes)

posterior_gamma_sd(alpha, beta, n, successes)
}
\arguments{
\item{population_avg}{the population average, a length 1 numeric value
describing the average (mean) value of the population}

\item{population_var}{the population variance, a length 1 numeric value
describing the variance of the population}

\item{alpha}{alpha parameter of a gamma distribution}

\item{beta}{beta parameter of a gamma distribution}

\item{n}{numeric value representing number of observations}

\item{successes}{numeric value representing the total number of successes across all the n observations.}
}
\description{
Functions for the gamma distribution to calculate priors and
posteriors.

The gamma distribution should be used as a rate per event, the distribution
being predicting is a value between 0 and inf. Think of this as runs/game
like in baseball.
}
\examples{

rpg_mean <- mean(lahman_team_runs_2010$runs_per_game)
rpg_var <- var(lahman_team_runs_2010$runs_per_game)

rpg_params <- get_gamma_parameters(population_avg = rpg_mean,
                                   population_var = rpg_var)

rpg_alpha_prior <- rpg_params$alpha
rpg_beta_prior <- rpg_params$beta

lahman_team_runs_2010 \%>\%
   mutate(
     posterior_rpg_avg = posterior_gamma_mu(rpg_alpha_prior, rpg_beta_posterior, G, R),
     posterior_rpg_sd = posterior_gamma_sd(rpg_alpha_prior, rpg_beta_posterior, G, R)
   )

}
\seealso{
\code{\link[=posterior_beta_mu]{posterior_beta_mu()}}, \code{\link[=posterior_beta_sd]{posterior_beta_sd()}} which is the beta distribution posterior calculations
}
